+++ common/StaticTools.cpp
@@ -8,8 +8,8 @@
 
 std::ofstream StaticTools::Log;
 
-uint32_t StaticTools::musicVolume = 100u;
-uint32_t StaticTools::soundVolume = 100u;
+float StaticTools::musicVolume = 100u;
+float StaticTools::soundVolume = 100u;
 
 std::pair<short, short> StaticTools::GetResolution(void)
 {
